requirements.txt
ollama
pytest
fastapi
uvicorn[standard]
sse-starlette
pydantic
requests
python-dateutil
regex
pydantic_settings
alembic
psycopg2-binary
python-dotenv
sqlalchemy